# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: Foxhound SERP Scraper API
  description: |
    REST API for the distributed SERP scraping + business-contact enrichment pipeline.

    Two parallel surfaces:
    - **v1** (`/api/*`) — bare-shape responses: `{...}` on success, `{"error":"message"}` on failure.
    - **v2** (`/api/v2/*`) — envelope responses: `{"data": ...}` on success, `{"error":{"code":"…","message":"…"}}` on failure.

    **Roles**: `admin` (full mutations) and `viewer` (GET-only). Authenticate with
    `Authorization: Bearer <token>` (from the login endpoints) or a static `x-api-key` header.

    **Geo conventions (v0.9.5+)**: listing `country` carries the legacy full name
    ("Indonesia"); `country_code` is ISO-3166-1 alpha-2 ("ID"); `geo_source` is empty for
    page-extracted geo and `query_inference` when inherited from the originating search query.
  version: 0.9.6
servers:
  - url: https://serpina.kremlit.dev
    description: Production (manager)
  - url: http://localhost:8080
    description: Local
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: v1 login — exchange username + api key for a bearer token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Token pair
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/V1Error" }
...
  /api/v2/auth/login:
    post:
      tags: [auth]
      summary: v2 login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Token pair (v2 envelope)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/V2Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token from /api/auth/login or /api/v2/auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  responses:
    V1Error:
      description: v1 bare error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
    V2Error:
      description: v2 envelope error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code: { type: string, example: internal_error }
                  message: { type: string }
  schemas:
    LoginRequest:
      type: object
      required: [username, api_key]
      properties:
        username: { type: string }
        api_key: { type: string }
...
    TokenResponse:
      type: object
      properties:
        token: { type: string }
        refresh_token: { type: string }
        expires_at: { type: string, format: date-time }
        role: { type: string, enum: [admin, viewer] }
...